This commit contains multiple independent changes. I'm sorry about that, but since I was working offline, I had no choice.
Introduced concept of a pesudochange event. If a model uses a toJSONDecorated method to provide calculated values, those values might depend on attributes of a related model or a collection of related models. Since changes in those attributes now effect the calculated values of the model, registered listeners need to be informed about them, to for example keep the model's view up to date. We can however not trigger a "change" event, as backbone-relational interferes with it's propagation. Therefore models that match the conditions described above are to send a "pseudochange" model instead. Event consumers should listen for the change event as well as the pseudochange event. 

Made AppView only rerender activeView when it is changed, instead of rerendering whole app.

Added a way to change expense.description to ExpenseEditView via input text field.

Renamed ParticipationView.toggleParticipating to ParticipationView.setParticipating as the method doesn't toggle but uses the checkbox value to determine Participation.participating